requirements.txt
numpy==2.2.5
scipy==1.15.2
h5py==3.14.0
pandas==2.2.3
matplotlib==3.10.1
scikit-learn==1.6.1
physics-tenpy==1.0.6
